refactor: moved EventOverview into EventOverviewScreen
