+++ packages/app/components/EventOverview/EventOverviewRegister.js
