+++ packages/app/screens/group/event_overview/EventOverviewScreen.js
@@ -1,23 +1,45 @@
 import React, { useEffect, useState } from "react";
-import { View, StyleSheet } from "react-native";
+import { View, StyleSheet, Image } from "react-native";
 import PropTypes from "prop-types";
 import axios from "axios";
 
-import EventOverview from "@app/components/EventOverview/EventOverview";
-import EventRegister from "@app/components/EventOverview/EventRegister";
+import EventOverviewText from "@app/components/EventOverview/EventOverviewText";
+import EventOverviewRegister from "@app/components/EventOverview/EventOverviewRegister";
+import FontAwesome5 from "@expo/vector-icons/FontAwesome5";
 import { ENDPOINT } from "@app/utils/constants";
 import * as SecureStore from "@app/utils/SecureStore";
 
-const screenColor = "white";
+const WHITE = "white";
+const transparentColor = "#00000077";
 
 const styles = StyleSheet.create({
+  backArrow: {
+    color: WHITE,
+    padding: 5.2,
+  },
+  backArrowView: {
+    backgroundColor: transparentColor,
+    borderRadius: 29 / 2,
+    height: 29,
+    left: 5,
+    position: "absolute",
+    resizeMode: "cover",
+    top: 5,
+    width: 29,
+  },
+  banner: {
+    height: 160,
+    resizeMode: "cover",
+    width: "100%",
+  },
   container: {
-    backgroundColor: screenColor,
+    backgroundColor: WHITE,
     flexDirection: "column",
     height: "100%",
   },
   overview: {
-    flex: 9,
+    flex: 8,
+    padding: 10,
   },
   register: {
     flex: 1,
@@ -55,33 +77,49 @@ export default function EventOverviewScreen({ navigation, route }) {
       hours -= 12;
       dayPeriod = "PM";
     }
+    if (minutes < 10) minutes = "0" + minutes;
 
-    if (minutes < 10) {
-      minutes = "0" + minutes;
-    }
-
-    return `${
-      date.getMonth() + 1
-    }/${date.getDate()}/${date.getFullYear()} at ${hours}:${minutes} ${dayPeriod}`;
+    return `${date.getMonth() + 1}/${date.getDate()}/${
+      date.getFullYear() - 2000
+    } at ${hours}:${minutes} ${dayPeriod}`;
   }
+
+  const onBackPress = () => {
+    navigation.navigate(previousScreen);
+  };
+
+  const BackArrow = () => {
+    return (
+      <View style={styles.backArrowView}>
+        <FontAwesome5
+          name="arrow-left"
+          size={18}
+          style={styles.backArrow}
+          onPress={onBackPress}
+        />
+      </View>
+    );
+  };
 
   return (
     <View style={styles.container}>
+      <Image
+        source={require("@app/assets/test_card_banner.png")}
+        style={styles.banner}
+      />
+      <BackArrow />
       <View style={styles.overview}>
-        <EventOverview
-          banner={require("@app/assets/test_card_banner.png")}
+        <EventOverviewText
           groupName="Software Engineering Association"
           title={event.title}
           timeBegin={formatDate(event.startDate)}
           timeEnd={formatDate(event.endDate)}
           location={event.location}
           description={event.description}
-          navigation={navigation}
-          previousScreen={previousScreen}
         />
       </View>
       <View style={styles.register}>
-        <EventRegister registerState={false} />
+        <EventOverviewRegister registerState={false} />
       </View>
     </View>
   );
